cloog_loop_merge: keep loops in original order if possible
cloog_loop_merge collects loops with intersecting domains.
Previously, it was possible for the order of the loops
within such a collection to be different from the original order.
Although this would only happen when the order is of no importance,
it could be confusing for a user to see that the statements are
generated in a different order for no apparent reason.

This patch keeps the loops within a collection
in the order of the list passed to cloog_loop_merge.