Don't increment statement->number when printing.
The rationale behind this incrementing is apparently that the
first statement would be printed as S1 rather than S0.
However, this also means that a user of CLooG as a library will
see the statement numbers she provides incremented for no good reason.

This patch moves the incrementing to cloog_loop_read.