configure: replace --with-polylib-{include,library} by --with-polylib-builddir
The -with-polylib-include and --with-polylib-library options, introduced
in b5fef7e7bf54802c428ecd19377e8aa4685790a2, were primarily meant
for building CLooG with PolyLib still in its build directory.  However,
this only worked when the source directory and the build directory
are the same, since the PolyLib header files need to be picked up
from both the source directory and the build directory.

The new --with-polylib-builddir specifies the build directory,
from which the source directory and the location of the
headers and libraries can be derived.

Note that the new option does not help for installations
in non-standard places.  If there really is such a need, the old
options can be reinstated for this purpose.

For consistency, This commit also renames the --with-polylib option
to --with-polylib-prefix.