configure: don't check polylib header file on --with-polylib-builddir
In commit 25bc1b2f2c50ec919c640575b14c2e0b1b482247 (configure.in:
don't require polylib to have been compiled prior to configure),
we allowed the use of a configured but not yet built polylib
to be specified, by not depending on the built libraries.
But the corresponding polylib header file is also not usable
yet, so we shouldn't depend on it either during configuration.