Test correctness of generated code, without requiring textual equality
The exact output of CLooG is not necessarily stable, in that
the order conditional tests are generated is not fixed.
Different versions of the PolyLib and a fortiori different backends
may result in non-identical, but equivalent code.
Furthermore, different backends or different versions of CLooG
may apply different code optimizations.
We therefore want to check whether the generated code is correct,
without relying on textual equality or manual inspection.

We do this by compiling and running code that executes both
a "known good" and a test version of the output for some values
of the parameters and checks whether both versions perform
the same operations for the same original loop iterations in the
same order.  This check itself is performed by having both
versions compute a hash value based on the iterators and
the statement numbers.

The original, textual comparison is kept, since we also
want to check whether we still perform the same optimizations
on the code.  However, these tests can be slightly different
for different backends, while the "correctness" check should
be shared among all backends.