Change semantics of cloog_domain_cut_first
cloog_domain_cut_first used to have very confusing action-at-a-distance
behavior, changing not only the domain it took as an argument, but
also any other domain that happened to reference the same CloogDomain
structure.

This patch makes cloog_domain_cut_first return two domains,
the "first" part of the union and the remainder of the union.
Any other domain that references the same CloogDomain structure
(if there is any) is left untouched.
All callers of cloog_domain_cut_first are changed accordingly.