Handle 0D problems by adding a 0th level to the main recursion
The main algorithm recurses over all dimensions of the loops,
the initial call working on the first, outermost dimension.
However, for 0-dimensional problems, there is no outermost
dimension and this initial call would therefore perform
all sorts of inappropriate actions resulting in general mayhem.

We solve this problem, by adding an extra 0th level in the recursion
that handles the parameter domains of the loops (which may themselves
be 0-dimensional).  Only some of the steps behave slightly differently
for the 0th level than they to for the other levels.

1. intersection with the context: proceeds as usual
2. projection: proceeds as usual; the loops are projected on the first
	0 dimension, i.e., onto the parameter space
3. separation: proceeds as usual, but we have to be careful about
	the "scalar" dimension optimization, which obviously does
	not apply to the parameter domain
4. lexographic sorting: not needed; the program is only run for
	a specific set of values for the parameters, so it does
	not matter in which order the disjoint parameter domains
	appear
5a. stride computation: not applicable
5b. recursion: proceeds as usual onto the first dimension
	(assuming there is a first dimension; otherwise the
	recursion correctly stops at the "0th" dimension)
6. backtrack: not needed

The clast generation has also been changed to expect this
extra level.  In particular, the 0th level does not generate
any loop, but it may generate a set of independent conditions.