cloog_loop_separate: exploit disjointness of loop domains
Before, cloog_loop_separate would only explit the disjointness
of loop domains if the disjointness was "obvious", i.e.,
if cloog_domain_lazy_disjoint returns true.
However, if we end up computing the intersection and that turns
out to be empty, then we know the two domains were disjoint as
well and we can also exploit this knowledge in this case.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>