Look for more opportunities for blocking statements after scattering
After scattering the domains and extracting the constant dimensions,
we can more easily check whether some statements can be blocked.
With these additionaly tests, the isl backend finds the same
blocks as those found by the PolyLib backend on the test examples.

Note that the PolyLib backend is a little bit too eager to find
blocks and gets it wrong sometimes.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>