clast: allow clast_term to represent multiple of any clast_expr
Before, a clast_term could only be used to represent a constant
or a multiple of a (const char *) name.  Now it can be used
to represent a multiple of any clast_expr, include the new
clast_name.

This functionality is not currently used, but will be useful
for the isl backend.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>