PolyLib backend: disable cloog_scattering_lazy_block
The original cloog_scattering_lazy_block is known to get it wrong
in some cases.  For example, on the newly add block.cloog test case,
it would make CLooG generate

    /* Generated from block.cloog by CLooG 0.14.0-170-g72daac3 64 bits in 0.01s. */
    S1();
    S2();
    S3(0);
    S3(1);

instead of the correct

    /* Generated from block.cloog by CLooG 0.14.0-170-g72daac3 gmp bits in 0.00s. */
    S1();
    S3(0);
    S2();
    S3(1);

Cedric Bastoul is working on a fix, but in the mean time it is safest
to disable the known incorrect implementation.  Once it has been fixed,
it can easily be resurrected.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>