move all global variables to CloogState
Putting all global variables in the CloogState ensures
that CLooG is thread safe as long as the backend used is
thread safe as well.

In particular, the PolyLib specifc MAX_RAYS is moved
into CloogState, even though it's probably not
very useful to have it set to different values in
different states.

Most of the global variables used in checking for memory
leaks have also been moved to CloogState.
The remaining two sets have been dropped, namely
cloog_int_* and cloog_matrix_*.
It would be tricky to pass the CloogState to all
init and clear macro calls, while there aren't that
many direct manipulations of cloog_int_t's left in
the core and there will be even fewer after one of
the next patches.
cloog_matrix is (now) a PolyLib specific thing.
It wouldn't be too difficult to have it keep a
CloogState reference as well, but it doesn't
seem worth it.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>