make CloogDomain and CloogScattering completely opaque types
Before, CloogDomain and CloogScattering were typedef'd to different
structures in the different backends.  Now they are typedef'd to
the same structure, but it is left incomplete in the backend independent
code, so there is no risk anymore of the backend independent code looking
into these structures.

There are currently only file and backend dependent interfaces
for creating these structures, but a backend independent interface
could easily be added.
The backend dependent interfaces are now also documented.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>