cloog_domain_stride: handle union domains
CLooG normally separates all domains into disjoint convex sets,
but cloog_domain_stride is called on domains that have not been
separated yet.  This commit adds support for union domains in
cloog_domain_stride.  In the case of the PolyLib backend, we
simply give up if we are handed a union domain.  It's not worth
it to provide a proper implementation for the PolyLib backend.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>