keep track of offsets of strided loops and use them when simplifying modulo guards
In insert_modulo_guard, a term a * i in the modulo expression,
with "i" a strided iterator and the stride of "i" a multiple of the modulo,
would simply get dropped.  This is only justified, if the offset
of the loop is zero (or at least a multiple of the modulo).
Instead, "a * i" should be replaced by "a * o" with "o" the offset
of the loop.

Problem reported by Johan Walters <jwalters@compaandesign.com>

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>