isl backend: detect equalities on all sets
Detecting equalities in isl used to be fairly expensive, but shouldn't be so
anymore.  This detection can sometimes significantly simplify the generated
code, especially in the presence of modulos.
There is some fallout on the darte example, but that will be fixed
when we move to a newer version of isl in the next commit.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>