cloog_loop_nest: also introduce a nesting for the deepest level
The regular operation of cloog_loop_generate constructs a nesting
for each level with the deepest nesting having the original loop
as an inner loop.  The clast construction depends on this property,
specifically when figuring out how many substitutions are needed
for a user statement.
cloog_loop_nest would not introduce a nesting for the deepest level,
resulting in missing arguments.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>