cloog_loop_generate: detect components and generate them separately
If two iteration domains overlap in such a way that you can execute
one completely before executing the other at a given level, then
the standard separation step will needlessly split off the shared iterations.
This patch detects such pairs of domains.  More precisely, it detects
clusters of domain such that one cluster is executed before the other.

A typical example is "non_optimal/youcef.cloog".
Before this patch, it would generate

    for (i=0;i<=3;i++) {
      S1(i,i);
      S2(i,i);
      for (j=i+1;j<=4;j++) {
	S2(i,j);
      }
      S2(i,5);
      S3(i,5);
    }
    S1(4,4);
    S2(4,4);
    S2(4,5);
    S3(4,5);
    S1(5,5);
    S2(5,5);
    S3(5,5);

With this patch, it generates

    for (i=0;i<=5;i++) {
      S1(i,i);
      for (j=i;j<=5;j++) {
	S2(i,j);
      }
      S3(i,5);
    }

There is some fall-out of this patch.

orc.cloog will now generate

    for (p1=2;p1<=3;p1++) {
      if (p1 == 2) {
	S4(0);
      }
      if (p1 == 3) {
	S1(1);
      }
    }

This happens during backtracking and is just an unlucky coincidence.
Similar code could get generated even without this patch.

walters.cloog will now generate

      div36 = floord(i,3);
      for (div37=max(ceild(i,3),ceild(2*i-3*div36-1,3));div37<=min(floord(i+2,3),floord(2*i-3*div36+1,3));div37++) {
	S4(i,div36,div37,i-div36-div37);
      }

This is the result of an unfortunate interaction with the
detection of equalities.  Arguably this is a user error.
The input file contains the following domain:

    8 6
      1  1  0  0  0 -1
      1 -1  0  0  0 10
      1  1 -3  0  0  0
      1 -1  3  0  0  2
      1  1  0 -3  0  2
      1 -1  0  3  0  0
      1  1  0  0 -3  1
      1 -1  0  0  3  1

The last three dimensions are not very useful.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>