move lower-bounds fix-up from cloog_loop_stride to clast generation
Updating the lower bound in cloog_loop_stride was a bit fragile
as it depends on the backend not changing the constraints in any
way later on.  If a backend fails to detect all stride information,
we could end up generating wrong code too.
By postponing the update to the point where we
actually write out the lower-bound, we avoid these problems.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>