make installed header files independent of configure options
We want to share most of the header files among the different
CLooG instantiations.  In particular, we don't want an install
of a non-gmp cloog-polylib to render any cloog-isl unusable,
or vice versa.

The down-side is that we require users of CLooG to specify
the integer type used inside the CLooG library they want to use
by defining one of CLOOG_INT_INT, CLOOG_INT_LONG, CLOOG_INT_LONG_LONG or
CLOOG_INT_GMP, but we will somewhat alleviate this in the next commit.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>