Turn off backtrack by default
The backtracking phase of the algorithm was added to eliminate redundant ifs,
but it comes at the cost of potentially serious code bloat.
With the decomposition into components and some recent improvements,
we should be able to remove most if not all of the redundant ifs
without the backtracking phase, so we can safely turn it off by default
and avoid the code bloat.

The -nobacktrack option is removed and replaced by a -backtrack option,
which turns on the backtracking.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>