isl backend: scattering relation now maps domain dimensions to time dimensions
The time/scattering dimensions used to be treated as input dimensions
of the scattering relation because they appear first in the PolyLib
input format.  However, in the extended PolyLib input format for
relations, it's the output dimensions that appear first, so we can
now consistently invert the mapping to obtain a more logical domain
and range.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>