clast.c: insert_guard: make sure enough space is allocated for guards
For regular dimensions at most two guards are generated,
one lower bound and one upper bound.  When level == 0,
however, parameter constraints are added individually, and
the number of constraints may obviously be more that twice
the number of parameters.  It's therefore not sufficient to
allocated space for "2 * total_dim" constraints.
Instead just allocated space for a number of guards that
is equal to the total number of constraints.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>