simplify cloog_loop_merge implementation
The original implementation would look at each loop in the list
in turn and decide whether it should be combined with earlier
loops in a convex hull or kept separately.  It's not clear
if this always works as the result may depend on the order
in which the loops are considered.
More importantly, since commit 49dbb13 (cloog_loop_generate: detect
components and generate them separately) there shouldn't be any
for checking whether or not a loop should be combined with another
loop.  If two loops should not be combined here, then they
would have ended up in different components.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>