cloog_loop_generate: optimize the loop domains after separation
The separation phase may split a domain into parts with only
a single inner loop.  In this case we want to replace the
iteration domain by the projection of the iteration domain
of the inner loop.
This can be seen as a light-weight backtracking phase.
When backtracking is turned on, the effect of this change is small,
but with backtracking turned off, it helps to recover some of the
benefits of backtracking.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>