dump ScatteringList in cloog_program_dump_cloog
In case a scattering list is supplied to cloog_program_dump_cloog
a CLooG input file is produced, that describes the CloogProgram (before the
scattering is applied) and the scattering to be applied to the CloogProgram.

In case no scattering list is supplied (scattering is NULL) the current
behaviour is unchanged. cloog_program_dump_cloog assumes the scattering was
already applied.

Signed-off-by: Tobias Grosser <grosser@fim.uni-passau.de>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>