allow library users to specify names of statements for use in pretty printing
The default pretty printer produces code that is very suitable for
use in interactive environments.  The only problem is that there is
no way of specifying the names of the statements.
It is of course possible for library users to write their own pretty
printer, but this would lead to a lot of duplication.
Instead, with very minor changes, the default pretty printer is changed
to take statement names into account, if provided.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>