detect strides even if the remainder of the lower bound depends on outer iterators
The original implementation of stride detection only works if the remainder
of the lower bound on division by a potential stride is a constant.
The new implementation allows this remainder to vary with each iteration
of the outer loop(s), while of course still remaining constant within
the same iteration of the outer loop(s).

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>