optionally store domains of scattering dimensions in clast user statements
It is sometimes useful to know for a given instance of a user statement
in a clast for which scattering dimension values it is executed.
This information is difficult if not impossible to retrieve from the
clast because some dimensions may have been removed if they attain
exactly one value.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>