cloog_union_domain_from_isl_union_set: actually use name of space, if any
Before this commit, an attempt was made to use the name of the space,
but the wrong tuple was being used.
Also, if the space is named, then it will get dropped by some operations
performed by CLooG, so we'd better drop the name completely, as well as
any internal structure, to avoid mismatches later on.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>