cloog_loop_simplify: don't break up domains with union contexts
Before, cloog_loop_simplify would simplify the intersection of
the original domain and the context with respect to the context.
It is, however, sufficient to simplify the original domain itself
with respect to the context.  Moreover, if the context is a union,
then the intersection could needlessly split up the domain into
several parts.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>