isl interface: remove name and internal structure of iteration domain spaces
CLooG has no use for them and also cannot handle them.
Names and internal structure usually disappear anyway due to some
of the operations performed on the sets, but in some special cases,
this would not happen soon enough, so let's just do it explicitly.
The new code also avoids the use of an undocumented and outdated isl function.

Reported-by: Tomo <tomofumi.yuki@gmail.com>
Tested-by: Tomo <tomofumi.yuki@gmail.com>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>