fix backtracking in presence of constant loops
9802965 (cloog_loop_generate_general: add special treatment for loops
with constant value, Sat Dec 4 17:57:26 2010 +0100) broke backtracking
because it would only retain the constraints on the constant loop iterator
during its special treatment of constant loops and not the constraints on
outer loop iterations.  This had gone mostly unnoticed because backtracking
is no longer the default.
We simply add back the constraints on the outer loops during backtracking.

Without this patch, the output of the new test case
(with options -f 1 -backtrack) is

    for (;;c1++) {
      if (c1 == 0) {
	S1(0);
      }
    }

Reported-by: Uday Kumar Reddy <udayreddy@gmail.com>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>