clast generation: don't feed malformed constraints to modulo guard construction
We try to simplify the constraints used during the construction
of a modulo guard.  In the isl backend, we end up calling isl_basic_set_gist,
but this may in some cases result in a constraint set that the caller
is unable to handle.  In particular, the result may contain more than
one existentially quantified variable and this would be treated in the
same way as when there were no existentially quantified variables left.
This then produced an error later on because the constraints are not in
the expected form.

If we end up with more than one existentially quantified variable,
we now just return the original set of constraints.

Reported-by: Louis-Noel Pouchet <pouchet@cse.ohio-state.edu>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>