fix check for whether user statement domains should be saved
5f6c710 (Optionally store domains of scattering dimensions in clast for,
Mon Nov 29 18:19:38 2010 -0500) introduced a test that would only
allow user statement domains to be saved if they entirely consisted
of scattering dimensions.  The test was wrong however.  In particular,
if code generation is stopped right after the last scattering dimension,
then the saved domain should be saved, but it was not because the nesting
level is one higher than allowed by the old test.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>