cloog_loop_constant: don't drop constraints on outer loop iterators
9802965 (cloog_loop_generate_general: add special treatment for loops
with constant value, Sat Dec 4 17:57:26 2010 +0100) improved code
generation in presence of the -noscalars option, but did so by creating
loops with domains that only had constraints on the current loop iterator
and not on any previous loop iterator.
This broke a fundamental assumption of the CLooG internals.

We already introduced a bandaid in b130274 (fix backtracking in presence
of constant loops, Thu Mar 24 21:48:27 2011 +0100) for backtracking,
but cloog_loop_stop was also affected.
Rather than introducing a new bandaid, let's just play by the rules
in cloog_loop_constant and remove the backtracking bandaid.
The fallout appears to be relatively minor.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>