include stride constraint in saved domains
When both the strides and the save_domains options are used,
the generated clast takes the strides into account, but the
associated domains would not contain any stride information.

Include the stride constraint so that the saved domain
matches the loop structure.
We need to do this in two places, one for the domains
on the clast_fors and one for those of clast_user_stmts.
For the former, we simply add the stride constraint right
before we save the domain.  For the latter, we add the stride
constraint to the context when recursing in the main code
generation algorithm.  If the current level is the last level
for which to generate code (according to the stop option),
then this context domain will end up as the domain of the
user statements and will therefore include the stride constraint.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>