clast: save domains on entire output tree
In 62da9f7 (optionally store domains of scattering dimensions in clast
user statements, Mon Sep 27 18:50:03 2010 +0200), domains of any dimensionality
would be saved.  In 5f6c710 (Optionally store domains of scattering
dimensions in clast for, Mon Nov 29 18:19:38 2010 -0500), a test was
added, fixed in a8cb7e1 (fix check for whether user statement domains
should be saved, Fri May 6 16:55:12 2011 +0200), to only save domains
consisting of scattering dimensions only.

Domains on the leaf nodes can therefore only be obtained if the
stop option is used, but by the very nature of this option, this
results in incomplete code.

Instead, we now save domains throughout, but project them on
the scattering dimensions if needed.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>