allow unrolling of loops with multiple lower bounds
When support for unrolling was added in 7f6868b (add support for
unrolling, Thu May 12 16:22:49 2011 +0200), it was restricted
to loops with only a single lower bound.
There is no need for this restriction, however.
Instead, it is sufficient that there be a lower bound
that can be incremented only a finite number of times.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>