configure.ac: use non-deprecated form of AM_INIT_AUTOMAKE
The old form was deprecated back in automake 1.6, so it's probably
about time to switch.
The new form allows us to specify automake options, which in turn
allows us to use "autoreconf -i" instead of the current autogen.sh script.

This is especially useful for OS X users.  OS X comes with its own
libtool, which is different from GNU's.  GNU libtool is therefore
usually made available as glibtool and libtoolize as glibtoolize.
The current autogen.sh script therefore does not work as is for those users.
autoreconf would have the same problem, except that the macport version
of autoconf has a pathed version of autoreconf which looks for
glibtoolize instead of libtoolize.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>