cloog_loop_free_parts: free loop->unsimplified
Since loop->unsimplified was introduced in 62da9f7 (optionally store
domains of scattering dimensions in clast user statements, Mon Sep 27
18:50:03 2010 +0200), it was only getting freed in cloog_loop_free
and not in cloog_loop_free_parts.

Now, loop->unsimplified is only set at the end of the computation,
so cloog_loop_free_parts is usually not called on a loop with
unsimplified set.  In fact, if it is, then this would mean that
the information about the unsimplied domain is lost.
Still, if we are losing the information anyway, we might as well
free it.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>