isl backend: represent equalities as isl_constraints
We used to represent the equalities in the CloogEqualities object
as isl_basic_sets.  Sometimes such a representation is convenient
because we need to perform operations on basic sets.  Some other
times, however, we would have to extract the equality constraint
from the isl_basic_set.  Since this latter operation is more
fragile than converting a constraint into an isl_basic_set,
it is much better to store the equalities as isl_constraints
instead and to convert them to isl_basic_sets when needed.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>