remove uses of cloog_constraint_clear from core
cloog_constraint_clear is not a very nice interface because it not
only affects the constraint itself, but also the constraint set
into which it points.  This forces a constraint to be tied to
a certain constraint set and this will no longer be the case
in future versions of isl.

There are two ways in which we avoid the use of cloog_constraint_clear.

The first is by deciding which constraints to not consider again based
on their properties instead of explicitly removing them.
The other is to use cloog_constraint_set_drop_constraint instead,
which also takes a pointer to the constraint set from which the
constraint should be removed, which allows for more independence
between constraint sets and constraints.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>