Only convert isl_sets to cloog_union_domains
Converting an isl_union_set containing more than one set is impossible,
because iterations from different sets are incomparable without
scattering. Instead of providing an interface that crashes with
an unrelated error message if more than one set is provided, we fix the
interface to take only a type that we can actually convert.

Signed-off-by: Tobias Grosser <tobias@grosser.es>