Add OpenScop support
This adds OpenScop support. The textual input is added through a
new -openscop option while the library input is added through a
new cloog_input_from_osl_scop function. For convenience, the
OpenScop Library is integrated to CLooG as a submodule. Some basic
tests are provided to check that the OpenScop input is not broken.
The OpenScop support is disabled by default.

Signed-off-by: Cedric Bastoul <cedric.bastoul@u-psud.fr>