Also report failures in special tests
The test suite run the tests in SPECIAL_TESTS, but it did not report
failing test cases in the final result. This was because the write to
'failedtest' was not propagated out of the surrounding shell loop. To
me it seems the 'read' command is resetting the variables to the state
before its execution. This patch fixes the failure by using 'sed' to
extract the program name as well as the relevant options.

Signed-off-by: Tobias Grosser <tobias@grosser.es>